--- a/src/ARMeilleure/Memory/ReservedRegion.cs
+++ b/src/ARMeilleure/Memory/ReservedRegion.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace ARMeilleure.Memory
 {
-    class ReservedRegion
+    public class ReservedRegion
     {
         public const int DefaultGranularity = 65536; // Mapping granularity in Windows.