--- a/src/Ryujinx.Cpu/LightningJit/AddressForm.cs
+++ b/src/Ryujinx.Cpu/LightningJit/AddressForm.cs
@@ -1,0 +1,18 @@
+
+namespace Ryujinx.Cpu.LightningJit
+{
+    enum AddressForm : byte
+    {
+        None,
+        OffsetReg,
+        PostIndexed,
+        PreIndexed,
+        SignedScaled,
+        UnsignedScaled,
+        BaseRegister,
+        BasePlusOffset,
+        Literal,
+        StructNoOffset,
+        StructPostIndexedReg,
+    }
+}