--- a/src/Ryujinx.Cpu/LightningJit/Arm32/BranchType.cs
+++ b/src/Ryujinx.Cpu/LightningJit/Arm32/BranchType.cs
@@ -1,0 +1,15 @@
+namespace Ryujinx.Cpu.LightningJit.Arm32
+{
+    enum BranchType
+    {
+        Branch,
+        Call,
+        IndirectBranch,
+        TableBranchByte,
+        TableBranchHalfword,
+        IndirectCall,
+        SyncPoint,
+        SoftwareInterrupt,
+        ReadCntpct,
+    }
+}