… src/Ryujinx.Cpu/LightningJit/Arm32/IInstEmit.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/ImmUtils.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/InstInfoForTable.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/InstName.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/InstTableA32.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/InstTableT16.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/InstTableT32.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/RegisterAllocator.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/RegisterUtils.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/Compiler.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmit.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitAbsDiff.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitAlu.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitBit.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitCommon.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitCrc32.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitDivide.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitExtension.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitFlow.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitGE.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitHalve.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitMemory.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitMove.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitMultiply.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitNeonArithmetic.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitNeonBit.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitNeonCommon.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitNeonCompare.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitNeonConvert.cs
… src/Ryujinx.Cpu/LightningJit/Arm32/Target/Arm64/InstEmitNeonCrypto.cs