### Overview

This is a new JIT compiler that is designed for Arm CPUs. The main reason for creating a new JIT is that the current one (ARMeilleure) can't easily take advantage of the fact that when translating Arm code to another Arm CPU, most code can remain untouched since its the same architecture.

The new JIT skips several steps that the current JIT would do, and as a result it can produce code much faster (and with even better quality).

It is important to note that **the new JIT currently does not support x86 host**, and there is no plan for x86 support. While I do think we can add support for other RISC architectures in the future, I think x86 is different enough from Arm that it would not be worth it, and the approach used by the current JIT is good enough.

### Motivation

Arm devices are becoming more popular, and we want to ensure we have good support for those devices in the future. Having a decent JIT That can target Arm CPUs is important for that. While it is always possible to improve the existing JIT, in order for it to be the best it can would require a completely different approach, that can take maximum advantage of the similarity between Arm and... well, Arm.

While on macOS, we can use hypervisor to run the guest code directly, there are other platforms where we don't have access to a hypervisor and would still depend on the JIT.

### How it works

On both the 32-bit and 64-bit modes, the original register allocation is preserved. This has several benefits:
- Easier debugging as the generated code can be compared to the original without the need for a mental remapping of the registers.
- Much faster compilation, register allocation is slow, even more so if you use an algorithm that produces good allocations.
- Generally lower code size and better code. The original binary most likely was compiled using a good register allocation with no time constraints, so it already has good allocation.

For the 32-bit mode, this is pretty simple since AArch64 has (almost) 32 general purpose registers, while AArch32 has (almost) 16. So we have double the amount of registers to work with. Registers W0-W14 are mapped to 32-bit registers R0-R14 directly. R15 is special (PC), it is also mapped to W15, but handled in a special way.

The 64-bit mode is more challenging, because the amount of general purpose registers is the same in this case. Some of those registers are special (SP, LR, FP, etc). We can't use special registers in translated guest code directly. So instead, if a function uses one of those special registers, it is remapped to regular registers that are not in use on that function. If there are no enough unused registers available on the function to remap all special registers, the function is truncated up to the point where the amount of registers reaches the limit. Truncated functions are basically split into smaller "sub-functions". At the end of the truncated block, it jumps to the next part of the function that has been translated separately. It would have been also possible to spill the registers instead, but it would be considerably more complex for little benefit. The number of truncated functions is very small, and is usually around 2%. Currently there is only one register that is reserved as temporary register. Increasing the amount of temporary registers it needs to reserve would also increase the amount of truncated functions.

Like ARMeilleure, this JIT is capable of translating entire functions (rather than a single basic block). However if may not be able to detect the end of a function perfectly, this is not really a problem, but can cause overlapping parts of the same function to appear on the JIT cache. It can also potentially have performance implications. The function block decoding is more limited than the one ARMeilleure, the new JIT does not extend functions backwards (if there is a backwards jump, from a loop for example), or forward jumps that are too far way. That said, it can handle most common cases, and can handle loops just fine. As long the function is well formed and doesn't have weird control flow, it can detect the entire function.

The guest state is loaded on the entry point of the function, and saved on all exit points. When another guest function is called, the guest state is saved to the context. When it returns, it compares the address of the next function it should execute (returned on X0), if it matches the next address it would execute on this function, it loads the state from context again and keeps executing, otherwise it returns, until it finds the right function or reaches the dispatcher. Like ARMeilleure, it performs data flow analysis to find the registers that are used and modified on each entry/exit point of the function. That way it only needs to save and restore from context what is really used. However, the results might not be optimal, this is not a simple problem to solve. The data flow analysis is also currently only available for Arm64 mode, Arm32 instead always saves and restores all registers that are used in the entire function. This is simpler and produces worse results. The reason is that I planned to skip saving/restoring the context entirely for Arm32 (as it can just keep the state in registers all the time). But in the end, I didn't do it and here we are. All register usage information is derived from the instruction table. The table says which registers each instruction uses, and the translator uses that to extract the register numbers from the instruction itself, and build register usage information. This is also used for register remapping. Since Arm32 is not doing data flow, I also did not include most of the register usage information in the instruction table, instead it only tracks register writes (which is required to know if a instruction changes the PC register, considered a jump).

Speaking about the instruction table, it uses an approach similar to ARMeilleure for finding the instruction from the opcode. One difference is the way how constraints are handled. On ARMeilleure, they are generally handled where its most convenient. Sometimes on the table, sometimes while decoding the instruction fields, something on the implementation itself. If necessary, it would call the `Undefined` instruction implementation, if a specific encoding is undefined. On the new JIT, all constraints are added to the table. To make this possible, each entry on the table has a blacklist of encodings that are not allowed. So if a specific encoding is "Undefined" or another instruction, it is added to that list. This approach allows us to know if an instruction is undefined up-front. It also allows filtering out instructions that requires specific extensions or ISA versions if we need.

### Performance improvements

I didn't really notice frame rate improvements on the few tests I did, but there are some noticeable improvements for startup time and code size. The measurements below have been done on a Ryzen 9 7900X CPU (with the JIT generating Arm64 code, obviously it can't actually execute it, being a x86 CPU).

The chart below shows the measurement of the time taken to compile all the game code in the main binary. The games used for the test are Shantae and the Pirate's Curse, Neptunia Game Maker R:Evolution and New Super Mario Bros U Deluxe.

![image](https://github.com/Ryujinx/Ryujinx/assets/5624669/6a3ca1a0-b8a3-4c70-b440-a4744b24122d)

Neptunia has the highest size among them, so naturally it was also the one that took most time to compile. The code size when uncompressed is 75 MB. Shantae code size is 2MB, and New Super Mario Bros U Deluxe 9 MB.

![image](https://github.com/Ryujinx/Ryujinx/assets/5624669/59b96ed3-70ff-4ba2-9922-b24b91e572d9)

Here we can see the code size increase ratio when compared to the original code. A ratio of 1x means that it has the same size as the original code, 2x would be double the size, etc. Generally, the new JIT can produce code that is between 4x or 5x the size of the original (in a few cases managing to be around the 3.5x mark), while ARMeilleure is between 7x and 8x with HighCq mode. ARMeilleure also has the LowCq mode, which trades code quality for compilation speed, and as a result, is faster but produces significantly worse code (usually being 16x or 17x larger than the original).

The takeaway is that the new JIT can produce better code, and do so much faster than the old JIT. The most dramatic difference is seen on Neptunia, which has the largest code size. It took 4 minutes and 26 seconds to compile with the old JIT on HighCq mode, while on the new JIT it took only 4.64 seconds. The old JIT produced 633MB of code in the HighCq mode, while the new one produced 348MB. This is almost half the size, while taking a fraction of the time.

For 32-bit games, the code size difference is not as large for a few reasons:
- As mentioned before, it is not doing data flow analysis for more precise register usage information, which means in many cases, it saves and restores registers that are not actually used on a given control flow path.
- 32-bit code does not map as naturally to Arm64 like 64-bit code does. That means it requires more instructions per guest instruction.

Regardless, it still manages to win against the old JIT on code size, even with those limitations. It is also much faster, with New Super Mario Bros U Deluxe taking 8.7 seconds to translate in LowCq mode, 17.3 in HighCq, and 3 seconds with the new JIT.

The much faster compilation means that we do not need PPTC at all with the new JIT, and can avoid all the problems that comes with it. It means no more disk space used for PPTC caches, no more issues with mods that change code, and now more slow-ish first time gameplay experience. To reiterate what I said earlier, those benefits are only for Arm CPUs, this JIT does not support x86.

There are visible improvements on boot times. Below we can see a comparison with Mario Kart 8 Deluxe.

Old JIT, PPTC disabled:

https://github.com/Ryujinx/Ryujinx/assets/5624669/7279a427-91d5-46e8-9b41-850b921987ab

Old JIT, PPTC enabled:

https://github.com/Ryujinx/Ryujinx/assets/5624669/27da9ad2-160d-4009-a2ca-285b73204644

New JIT:

https://github.com/Ryujinx/Ryujinx/assets/5624669/db8a83f4-f10e-4225-b9af-ce0a6dbdc17b

Not surprisingly, the new JIT can beat the old one with PPTC disabled. Even with PPTC enabled, it manages to be a little bit faster (by roughly 2 seconds). One thing I did notice is that the second boot seems to be noticeably faster than the first one with the new JIT. I did not have time to investigate why yet, but it might be the cost of the .NET JIT. For all tests above, I ran the game at least once beforehand to ensure .NET JIT compilation time is not showing.

### Compatibility improvements

There are some games that did not work with the old JIT on Arm. They did work fine with hypervisor on macOS, so it's not a big deal, but it could be problematic for other platforms if we used JIT for them.

Luigi's Mansion 3, which would crash before showing the title screen:
<img width="1392" alt="Captura de Tela 2023-12-24 às 02 43 06" src="https://github.com/Ryujinx/Ryujinx/assets/5624669/9d8abb9b-f581-4f39-a4eb-91c51da6c411">
Pokémon Scarlet, which would softlock before getting in-game:
<img width="1392" alt="Captura de Tela 2023-12-24 às 02 42 26" src="https://github.com/Ryujinx/Ryujinx/assets/5624669/eef8ef0b-6e7e-46bf-aade-e8a19b52d5c2">

### Future improvements

Both Arm32 and Arm64 can be improved in the future:
- Investigate improvements to register state load/store.
- Re-use pointers for memory access when the guest register is the same.
- Some peephole optimization opportunities on Arm32.
- Endian flag is currently ignored on Arm32. We should implement it in the future, but no game actually uses this. The old JIT does not actually use it in all places either (for example, it is missing for VLDn/VSTn instructions, and also VLDM/VSTM iirc).

While working on this, I also found some problems on the old JIT:
- NEON comparison instructions (VCGE_Z and friends) ignores size field, and assumes that floating point size is FP32 even when it is FP16. Not a big deal since the Switch CPU does not support FP16, but it should be an Undefined instruction.
- VMAXNM and friends ignores sz field. Just like the above.
- There is a difference of 1ulp for VRECPS/VRSQRTS results compared to the native instruction on Arm64. Not sure what is up with that...
- VTBL and VTBX appears produce wrong results if the table and destination registers are the same. Weirdly enough, it seems that this is intentional (see https://github.com/Ryujinx/Ryujinx/blob/master/src/ARMeilleure/Instructions/InstEmitSimdMove32.cs#L294C25-L294C41). x86 has a SSSE3 path that might not have this issue, but Arm has no fast path.
- VLD2_1/VST2_1 with size = 0 and index_align<0> = 1. The register increment value is incorrect in this case, but I was not able to test it against Unicorn (it just freezes when I add this test case, for some reason...)

### Alternative approaches

Switch emulators on Android have been using an approach that they call "NCE" (Native Code Execution). It can run the guest code (almost) directly without hypervisor, but it comes with limitations (or rather, the need to hack things around to make it work):

- We can't run system instructions such as SVC, MRS, MSR etc directly, so patching the code to make it call the emulator is required. Those modifications are visible to the game.
- Interrupting the guest threads becomes somewhat complicated, since we can't insert "interruption points" in the code. We can use `pthread_kill` on Unix-like OS, but Windows has no such a thing.
- Worse by far, the code will access the emulator address space directly, so we need to set it up in a way that makes the game happy.

The last one deserves deserves its own section with the amount of problems it introduces:

- On many devices, the address space size is 39-bit. The Switch also has 39-bit address space, which means its not possible to fit it within the emulator address space. We need to shrink the guest address space to make it fit, which has the potential to break games as we differ from the Switch. This problem also affects JIT when using the host mapped memory manager modes.
- The allocated guest region must be contained entirely inside the guest address space. This means that 36-bit games won't work, as we most likely won't be able to allocate 36-bit worth of virtual memory right at the start of the emulator address space. There are not many of those, so it is not considered a big deal.
- For platforms with 16KB page size, there are additional challenges as we can't map the text segment and data segments as RX and RW respectively, since they are 4KB aligned, not 16KB aligned. We also can't just map them all as RWX if the platform has W^X. Making this work requires more hacks.

That said, we could consider implementing it in the future.

### Testing

Since this is a completely new JIT, there is a lot of things that have the potential to be broken, so testing is welcome. Note that it should be compared to the old JIT (hypervisor disabled). If the issue also happens with the old JIT, it is not a regression and it's probably not JIT related. 32-bit game testing is more important since we can't use hypervisor for that, so problems there would be more visible.

Note that the new JIT does not support the software memory manager mode, so it should be tested with the host mapped mode. Using the software mode will still use the old JIT.