@piplup55 2023-12-24
Nintendo 64 Apps now start on firmware 13, firmware 14 and above still crashes
![image](https://github.com/Ryujinx/Ryujinx/assets/100526773/9072beb3-00d0-479b-8a1c-525a7a2c3e80)
[Ryujinx_1.1.0+b264f3e_2023-12-24_17-05-49.log](https://github.com/Ryujinx/Ryujinx/files/13762421/Ryujinx_1.1.0%2Bb264f3e_2023-12-24_17-05-49.log)
[Ryujinx_1.1.1100_2023-12-24_17-09-30.log](https://github.com/Ryujinx/Ryujinx/files/13762422/Ryujinx_1.1.1100_2023-12-24_17-09-30.log)
related issue https://github.com/Ryujinx/Ryujinx/issues/6021