@merryhime 2023-12-24
> There is a difference of 1ulp for VRECPS/VRSQRTS results compared to the native instruction on Arm64. Not sure what is up with that...

This is because ARMeilleure's implementation of these instructions don't fuse mul-add.