@MutantAura 2023-12-27
Tears of the Kingdom seems to be crashing on boot with the new JIT.
[Ryujinx_1.1.0+45478f4_2023-12-27_13-10-25.log](https://github.com/Ryujinx/Ryujinx/files/13778777/Ryujinx_1.1.0%2B45478f4_2023-12-27_13-10-25.log)

Other titles tested include...

BoTW:
Old JIT - 36FPS, 10s boot, 9W (@30fps)
New JIT - 40FPS, 4s boot, 7.7W (@30fps)
HV - 49FPS, 2s boot, 7.3W (@30fps)

Red Dead Redemption:
Old JIT - 30FPS, 3s boot, 15.2W (@30fps)
New JIT - 53FPS, 2.5s boot, 8.7W (@30fps)
HV - 56FPS, <1s boot, 8.3W (@30fps)

Pokemon Mystery Dungeon: Rescue Team DX:
Old JIT - 47FPS, 8s boot, 7.9W (@30fps)
New JIT - 64FPS, 1.5s boot (fine), 4.8W (@30fps)
HV - 67FPS, <1s boot (inconsistent), 4.6W (@30fps)

Mario Kart 8 Deluxe:
Old JIT - 60FPS cap, 9s boot, 8.7W (@60fps)
New JIT - 60FPS cap, 1.5s boot, 7.7W (@60fps)