@gdkchan 2023-12-27
> Tears of the Kingdom seems to be crashing on boot with the new JIT.
[Ryujinx_1.1.0+45478f4_2023-12-27_13-10-25.log](https://github.com/Ryujinx/Ryujinx/files/13778777/Ryujinx_1.1.0%2B45478f4_2023-12-27_13-10-25.log)
Ignore me, old JIT does this too. Not sure if it's easier to debug with the new one?

IIRC its due to the hacks to make host mapped memory manager mode work wth 16KB pages.