@BoofOof32 2023-12-29
I have a [Thinkpad X13s](https://www.lenovo.com/us/en/p/laptops/thinkpad/thinkpadx/thinkpad--x13s-(13-inch-snapdragon)/len101t0019), one of the rare few Windows on ARM laptops that can reasonably run Linux.
I've tested Ryujinx on this device before, and got varying results 

On Vulkan with PPTC disabled on both master and PR build
```
SoC: Snapdragon 8cx Gen 3
RAM: 32GB LPDDR4x 4266MHz
GPU: Adreno 690
Driver: Mesa 23.2.1-ubuntu3.1
Distro: Ubuntu 23.10
```

- Good Job!
Old JIT - 30s boot time
New JIT - 17s boot time

- Mario Kart 8 Deluxe
Old JIT - 22s boot time
New JIT - 12s boot time

- Mario Party Superstars
Old JIT - perpetually loads
New JIT - 13s boot time

- Mario + Rabbids Kingdom Battle
Old JIT - 41s boot time (22-28fps)
New JIT - 25s boot time (roughly same perf)

- Metroid Prime Remastered
Old JIT - 12s boot time
New JIT - 10s boot time

- Red Dead Redemption
Old JIT - 42s boot time (4fps)
New JIT - 36s boot time

- Splatoon 2
Old JIT - 21s boot time
New JIT -18s boot time

- Super Mario 3D World + Bowser's Fury
Old JIT - 9s boot time
New JIT - 5s boot time

- Super Mario Odyssey
Old JIT - 20s boot time
New JIT - 17s boot time

- Super Mario Party
Old JIT - 32s boot time (40-60fps)
New JIT - 13s boot time (60fps)

- Super Smash Bros. Ultimate
Old JIT - 28s boot time
New JIT - 11s boot time

- Tears of The Kingdom
- Old JIT - 42s boot time
New JIT - 17s boot time

(I measure boot time as time between clicking a game and getting to the menu, so if the time looks long, it's probably because the intro is long)

This is not accounting for the poor performance or major graphical errors present on this device on most titles (either Turnip not knowing what to do with Ryujinx, or Ryujinx just not being optimized for Adreno, or a mix of both). The JIT doesn't  really affect performance apart from maybe 1 or 2 games of mine. but the difference is felt even on this random laptop